renamed conv_data_fmt -> convert_sh_fmt and also updated init
