+++ pyshbundle/__init__.py
@@ -73,7 +73,7 @@ __all__ = ['basin_avg',
            'sc2cs',
            'tws_cal',
            'visualisation_utils',
-           'conv_data_fmt']
+           'convert_sh_fmt']
 
 #Import individual modules
 from .basin_avg import BasinAvg
@@ -105,7 +105,7 @@ from .reader_replacer_csr import reader_replacer_csr
 from .reader_replacer_itsg import reader_replacer_itsg 
 from .reader_replacer_jpl import reader_replacer_jpl
 #from .sc2cs import sc2cs
-from .conv_data_fmt import sc2cs, clm2cs, clm2sc, cs2sc, klm2sc
+from .convert_sh_fmt import sc2cs, clm2cs, clm2sc, cs2sc, klm2sc
 from .tws_cal import TWSCalc
 # from .tws_py import 
 from .visualisation_utils import sc_triplot, cs_sqplot, polar_plot, mapfield, ylm, ylm_plot, gshs_prepare
